+++ tds/src/main/java/thredds/server/ncss/view/dsg/point/PointSubsetWriterXML.java
@@ -34,8 +34,8 @@ package thredds.server.ncss.view.dsg.point;
 
 import org.springframework.http.HttpHeaders;
 import thredds.server.ncss.exception.NcssException;
-import thredds.server.ncss.controller.NcssRequestUtils;
 import thredds.util.ContentType;
+import thredds.util.TdsPathUtils;
 import ucar.ma2.Array;
 import ucar.ma2.StructureData;
 import ucar.nc2.VariableSimpleIF;
@@ -73,7 +73,7 @@ public class PointSubsetWriterXML extends AbstractPointSubsetWriter {
 
         if (!isStream) {
             httpHeaders.set("Content-Location", datasetPath);
-            String fileName = NcssRequestUtils.getFileNameForResponse(datasetPath, ".xml");
+            String fileName = TdsPathUtils.getFileNameForResponse(datasetPath, ".xml");
             httpHeaders.set("Content-Disposition", "attachment; filename=\"" + fileName + "\"");
         }
